优化

+++ src/Controller/Passport.php
@@ -14,7 +14,7 @@ use Laket\Admin\Facade\Admin;
 use Laket\Admin\Support\Screen;
 
 /**
- * 登陆
+ * 登录
  *
  * @create 2021-3-18
  * @author deatil
@@ -30,12 +30,12 @@ class Passport extends Base
     }
     
     /**
-     * 登陆
+     * 登录
      */
     public function login()
     {
         if (Admin::isLogin()) {
-            $this->error("你已经登陆", laket_route("admin.index.index"));
+            return $this->error("你已经登录", laket_route("admin.index.index"));
         }
         
         // 使用 RSA 方法
@@ -68,12 +68,12 @@ class Passport extends Base
     }
     
     /**
-     * 提交登陆
+     * 提交登录
      */
     public function loginCheck()
     {
         if (Admin::isLogin()) {
-            return $this->error('你已经登陆！');
+            return $this->error('你已经登录！');
         }
         
         $verify = request()->post('verify');
@@ -87,14 +87,14 @@ class Passport extends Base
         
         // 验证数据
         $rule = [
-            'name|用户名' => 'require|alphaDash|length:3,20',
+            'name|用户名'   => 'require|alphaDash|length:3,20',
             'password|密码' => 'require',
         ];
         $message = [
-            'name.require' => '用户名不能为空',
-            'name.alphaDash' => '用户名格式错误',
-            'name.length' => '用户名字符长度错误',
-            'password.require' => '密码不能为空',
+            'name.require'     => '用户名不能为空！',
+            'name.alphaDash'   => '用户名格式错误！',
+            'name.length'      => '用户名字符长度错误！',
+            'password.require' => '密码不能为空！',
         ];
         $result = $this->validate($data, $rule, $message);
         if (true !== $result) {
@@ -104,7 +104,7 @@ class Passport extends Base
         // 密码
         $password = base64_decode($data['password']);
         if (empty($password)) {
-            return $this->error("用户名或者密码错误");
+            return $this->error('用户名或者密码错误！');
         }
 
         try {
@@ -119,17 +119,23 @@ class Passport extends Base
             $password = $rsakey->withPadding(RSA::ENCRYPTION_PKCS1)
                 ->decrypt($password);
         } catch(\Exception $e) {
-            return $this->error("用户名或者密码错误");
+            return $this->error('用户名或者密码错误！');
         }
 
-        if (! Admin::login($data['name'], $password)) {
-            $this->error("用户名或者密码错误", laket_route("admin.passport.login"));
+        $adminInfo = Admin::login($data['name'], $password);
+        if (empty($adminInfo)) {
+            return $this->error('用户名或者密码错误！');
+        }
+        
+        // 是否锁定
+        if ($adminInfo['status'] != 1) {
+            return $this->error('您的帐号已被锁定！');
         }
         
         // 清除数据
         Session::delete($prikeyCacheKey);
         
-        $this->success('登陆成功', laket_route('admin.index.index'));
+        return $this->success('登录成功！', laket_route('admin.index.index'));
     }
     
     /**
@@ -138,12 +144,14 @@ class Passport extends Base
     public function logout()
     {
         if (! Admin::isLogin()) {
-            $this->error("你还没有登陆", laket_route("admin.passport.login"));
+            return $this->error("你还没有登录", laket_route("admin.passport.login"));
         }
         
         if (Admin::logout()) {
-            $this->success('退出成功', laket_route("admin.passport.login"));
+            return $this->success('退出成功', laket_route("admin.passport.login"));
         }
+        
+        return $this->error("退出失败", laket_route("admin.passport.login"));
     }
     
     /**
@@ -155,7 +163,7 @@ class Passport extends Base
         
         make(Screen::class)->lock($url);
         
-        $this->success('屏幕锁定成功');
+        return $this->success('屏幕锁定成功');
     }
     
     /**
@@ -167,12 +175,12 @@ class Passport extends Base
         $password = request()->post('password');
         
         if (!Admin::checkPassword($adminInfo['name'], $password)) {
-            $this->error("密码错误，解除锁定失败");
+            return $this->error("密码错误，解除锁定失败");
         }
         
         make(Screen::class)->unlock();
         
-        $this->success('屏幕解除锁定成功');
+        return $this->success('屏幕解除锁定成功');
     }
 
 }
